crewai_orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cybervpn-crewai-orchestrator"
version = "0.1.0"
description = "CrewAI multi-agent orchestrator for CyberVPN - parallel to Claude Code agents"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=1.8.0,<2.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
